+++ game-frontend/projects/game-admin/src/app/components/faction-crud/faction-crud.component.ts
@@ -1,16 +1,13 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 
+import { ModalComponent } from '@owge/core';
 import { Faction, FactionUnitType } from '@owge/faction';
-import { ImageStore, UnitType } from '@owge/universe';
+import { ImageStore } from '@owge/universe';
 import { take } from 'rxjs/operators';
 
 import { AdminFactionService } from '../../services/admin-faction.service';
 import { AdminUnitTypeService } from '../../services/admin-unit-type.service';
-
-interface UnitTypeWithOverrides extends UnitType {
-  overrideId?: number;
-  overrideMaxCount?: number;
-}
+import { UnitTypeWithOverrides } from '../../types/unit-type-with-overrides.type';
 
 @Component({
   selector: 'app-faction-crud',
@@ -18,6 +15,8 @@ interface UnitTypeWithOverrides extends UnitType {
   styleUrls: ['./faction-crud.component.scss']
 })
 export class FactionCrudComponent {
+
+  @ViewChild(ModalComponent) public unitTypesOverridesModal: ModalComponent;
 
   public selectedEl: Faction;
   public unitTypes: UnitTypeWithOverrides[];
@@ -80,14 +79,27 @@ export class FactionCrudComponent {
   }
 
   public async onSelected(faction: Faction): Promise<void> {
+    this.selectedEl = faction;
     this.unitTypes = await this._adminUnitTypeService.findAll().pipe(take(1)).toPromise();
     const overrides: FactionUnitType[] = await this.adminFactionService.findUnitTypes(faction.id).toPromise();
     overrides.forEach(override => {
       const unitType: UnitTypeWithOverrides = this.unitTypes.find(current => current.id === override.unitTypeId);
       if (unitType) {
+        unitType.isOverride = true;
         unitType.overrideId = override.id;
         unitType.overrideMaxCount = override.maxCount;
       }
     });
+  }
+
+  /**
+   *
+   *
+   * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+   * @since 0.10.0
+   */
+  public saveOverrides(): void {
+    const overrideUnitTypes: UnitTypeWithOverrides[] = this.unitTypes.filter(current => current.isOverride);
+    this.adminFactionService.saveUnitTypes(this.selectedEl.id, overrideUnitTypes).subscribe(() => this.unitTypesOverridesModal.hide());
   }
 }