+++ game-frontend/projects/game-admin/src/app/types/unit-type-with-overrides.type.ts
@@ -1,0 +1,14 @@
+import { UnitType } from '@owge/core';
+
+/**
+ *
+ *
+ * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+ * @since 0.10.0
+ * @export
+ */
+export interface UnitTypeWithOverrides extends UnitType {
+    overrideId?: number;
+    overrideMaxCount?: number;
+    isOverride?: boolean;
+}